requirements.txt
streamlit
folium
streamlit-folium>=0.13.0
geopandas
pandas
numpy
plotly
scikit-learn
prophet
openai
pdfplumber
python-dotenv
joblib
requests
pyarrow
branca
shapely
fiona